+++ ci/install-os-dependencies.sh
@@ -1,10 +1,13 @@
 #!/usr/bin/env bash
 
+set -euo pipefail
+
 case "$(uname -s)" in
   Darwin)
-    brew install dpkg fakeroot
+    brew install dpkg
     ;;
   Linux)
+    sudo apt-get --quiet update
     sudo apt-get --yes --quiet install lintian
     ;;
 esac