+++ src/spawn.js
@@ -1,23 +1,15 @@
-'use strict'
-
-const { spawn } = require('@malept/cross-spawn-promise')
+import { spawn as crossSpawn } from '@malept/cross-spawn-promise'
 
 function updateExecutableMissingException (err, hasLogger) {
-  if (hasLogger && err.code === 'ENOENT') {
-    const isFakeroot = err.syscall === 'spawn fakeroot'
-    const isDpkg = !isFakeroot && err.syscall === 'spawn dpkg'
+  if (hasLogger && err.code === 'ENOENT' && err.syscall === 'spawn dpkg-deb') {
+    const installer = process.platform === 'darwin' ? 'brew' : 'apt-get'
 
-    if (isFakeroot || isDpkg) {
-      const installer = process.platform === 'darwin' ? 'brew' : 'apt-get'
-      const pkg = isFakeroot ? 'fakeroot' : 'dpkg'
-
-      err.message = `Your system is missing the ${pkg} package. Try, e.g. '${installer} install ${pkg}'`
-    }
+    err.message = `Your system is missing the dpkg package. Try, e.g. '${installer} install dpkg'`
   }
 }
 
-module.exports = async function (cmd, args, logger) {
-  return spawn(cmd, args, {
+export default async function spawn (cmd, args, logger) {
+  return crossSpawn(cmd, args, {
     logger,
     updateErrorCallback: updateExecutableMissingException
   })