+++ test/spawn.js
@@ -1,7 +1,7 @@
-'use strict'
+import { after, before, describe, it } from 'node:test'
+import { expect } from 'chai'
 
-const chai = require('chai')
-const spawn = require('../src/spawn')
+import spawn from '../src/spawn.js'
 
 describe('spawn', () => {
   let oldPath
@@ -11,12 +11,12 @@ describe('spawn', () => {
     process.env.PATH = '/non-existent-path'
   })
 
-  it('should throw a human-friendly error when it cannot find dpkg or fakeroot', async () => {
+  it('should throw a human-friendly error when it cannot find dpkg-deb', async () => {
     try {
-      await spawn('dpkg', ['--version'], msg => {})
-      throw new Error('dpkg should not have been executed')
+      await spawn('dpkg-deb', ['--version'], () => {})
+      throw new Error('dpkg-deb should not have been executed')
     } catch (error) {
-      chai.expect(error.message).to.match(/Error executing command \(dpkg --version\):\nYour system is missing the dpkg package/)
+      expect(error.message).to.match(/Error executing command \(dpkg-deb --version\):\nYour system is missing the dpkg package/)
     }
   })